TIP-1287: move ComputeCompletenessOnFamilyUpdateSubscriber
