+++ src/Akeneo/Pim/Enrichment/Bundle/EventSubscriber/Family/ComputeCompletenessOnFamilyUpdateSubscriber.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Pim\Enrichment\Bundle\EventSubscriber;
+namespace Akeneo\Pim\Enrichment\Bundle\EventSubscriber\Family;
 
 use Akeneo\Pim\Enrichment\Bundle\Doctrine\ORM\Query\FindAttributesForFamily;
 use Akeneo\Pim\Structure\Component\Model\AttributeInterface;
